Click on a beer from beer list to display beer data

+++ app/src/main/java/com/example/brewfinder/data/beerdata/Beer.kt
@@ -15,7 +15,7 @@ data class Beer (
     @SerializedName("beer_abv")
     val beerAbv: Double,
     @SerializedName("beer_ibu")
-    val beerIbu: Double,
+    val beerIbu: Int,
     @SerializedName("beer_description")
     val description: String,
     @SerializedName("rating_score")
